feat: add new button component for the note card
