+++ frontend/src/components/NotesDisplayCard.jsx
@@ -2,8 +2,7 @@ import { useState, useEffect } from "react";
 import styles from "../styling/NotesDisplayCard.module.scss";
 import Note from "./Note";
 
-export default function NotesDisplayCard({ group, user }) {
-  const [open, setOpen] = useState(false);
+export default function NotesDisplayCard({ group, user, handleClose }) {
   const [notes, setNotes] = useState([]);
   async function loadNotes() {
     console.log("group", group);
@@ -18,14 +17,6 @@ export default function NotesDisplayCard({ group, user }) {
     loadNotes();
     console.log("user", user);
   }, []);
-
-  const handleOpen = () => {
-    setOpen(true);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
 
   // const handleCreate = () => {
   //   // create a new note (for testing purposes)
@@ -46,39 +37,28 @@ export default function NotesDisplayCard({ group, user }) {
 
   return (
     <>
-      <button
-        type="button"
-        className={styles.notesButton}
-        onClick={() => {
-          handleOpen();
-        }}
-      >
-        Notes
-      </button>
-      {open && (
-        <div>
-          <div
-            className={styles.overlay}
-            role="button"
-            tabIndex={0}
-            onClick={handleClose}
-            onKeyDown={handleKeyPress}
-            aria-label="Close Card"
-          />
+      <div>
+        <div
+          className={styles.overlay}
+          role="button"
+          tabIndex={0}
+          onClick={handleClose}
+          onKeyDown={handleKeyPress}
+          aria-label="Close Card"
+        />
 
-          <div className={styles.noteCard}>
-            {notes.map((note) => (
-              <Note
-                key={note.id}
-                role={note.role}
-                id={note.id}
-                group={group}
-                user={user}
-              />
-            ))}
-          </div>
+        <div className={styles.noteCard}>
+          {notes.map((note) => (
+            <Note
+              key={note.id}
+              role={note.role}
+              id={note.id}
+              group={group}
+              user={user}
+            />
+          ))}
         </div>
-      )}
+      </div>
     </>
   );
 }
