+++ frontend/src/components/PlayPageNoteButton.jsx
@@ -1,0 +1,15 @@
+import styles from "../styling/NotesDisplayCard.module.scss";
+
+export default function PlayPageNoteButton({ handleOpen }) {
+  return (
+    <button
+      type="button"
+      className={styles.notesButton}
+      onClick={() => {
+        handleOpen();
+      }}
+    >
+      Notes
+    </button>
+  );
+}
